chatbot/requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-google-genai
pydantic
python-dotenv
streamlit
sentence-transformers
faiss-cpu
streamlit
langchain_huggingface
pypdf
wikipedia
langchain_openai
arxiv
langchain.classic
langgraph
langchain-openai
langchain_mistralai
pyarrow>=14.0.0